backend/requirements.txt
asgiref==3.8.1
astroid==3.2.4
bcrypt==4.2.0
cffi==1.17.0
colorama==0.4.6
coverage==7.6.1
cryptography==43.0.0
dill==0.3.8
Django==5.1
django-redis==5.4.0
djangorestframework==3.15.2
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg==3.2.1
pycparser==2.22
PyJWT==2.9.0
pylint==3.2.6
pytest==8.3.2
pytest-cov==5.0.0
redis==5.0.8
setuptools==72.1.0
sqlparse==0.5.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.1
websockets==12.0
wheel==0.43.0
